Parse errors from deps now honour the users chosen output
